Updated config.g
Added "M501 ;  Read stored parameters" as well as a description and example for users that will question what it is and why it's there.

https://duet3d.dozuki.com/Wiki/GCode#Section_M501_Read_stored_parameters

I use it in my own config.g for all my machines that run dc42's fork of reprapfirmware. I highly recommend this being a default, it is exceptionally useful. This line must always be the last line in the config.g